fix(ui): rewrite views w/o stale
