--- packages/ui/src/views/About.vue
@@ -1,4 +1,0 @@
-<template lang="pug">
-  div.about
-    h1 This is an about page
-</template>
